refactor(event): separate the parsing in modules
Separate the parsing of the DeviceEvent in different modules.
